Converted create house section to independent file
