+++ client/src/components/ViewHouse.tsx
@@ -10,6 +10,7 @@ import Leave from "@mui/icons-material/ExitToApp";
 import VerifiedIcon from "@mui/icons-material/Verified";
 import DoNotDisturbIcon from "@mui/icons-material/DoNotDisturb";
 import "./viewHouse.scss";
+import CreateHouse from "./CreateHouse";
 
 interface Values {
   house_name: String;
@@ -83,37 +84,6 @@ function ViewHouse() {
   const handleHouseChange = (event: SelectChangeEvent) => {
     setHouse(event.target.value as string);
   };
-
-  const formik_create = useFormik({
-    initialValues: {
-      name: "",
-    },
-    validationSchema: Yup.object({
-      name: Yup.string().max(64, "Name too long.").required(),
-    }),
-
-    onSubmit: (values) => {
-      fetch("/house/add", {
-        method: "post",
-        headers: {
-          Accept: "application/json",
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          name: values.name,
-        }),
-      }).then((response) => {
-        if (response.redirected) {
-          window.location.href = response.url;
-        } else {
-          response.json().then((responseJson) => {
-            console.log(responseJson["issue"]);
-            setIssue(responseJson["issue"]);
-          });
-        }
-      });
-    },
-  });
 
   const handleDeleteClick = () => {
     fetch("/house/delete").then((response) => {
@@ -281,46 +251,7 @@ function ViewHouse() {
                   </form>
                 </div>
               )}
-              {!joinHouse && (
-                <div className="house create">
-                  <form onSubmit={formik_create.handleSubmit}>
-                    <div className="form">
-                      <p className={"formIssue " + (issue && "active")}>
-                        {issue}
-                      </p>
-                      <div className="fields">
-                        <p
-                          className={
-                            "issue " +
-                            (formik_create.touched.name &&
-                              formik_create.errors.name &&
-                              "active")
-                          }
-                        >
-                          {formik_create.errors.name}
-                        </p>
-                        <div className="field">
-                          <label className={"label text"}>{"Name"}</label>
-                          <input
-                            className={"input text"}
-                            type={"text"}
-                            name={"name"}
-                            value={formik_create.values.name}
-                            onChange={formik_create.handleChange}
-                            onBlur={formik_create.handleBlur}
-                          />
-                        </div>
-
-                        <div className="button">
-                          <button disabled={!formik_create.isValid}>
-                            <span>Create</span>
-                          </button>
-                        </div>
-                      </div>
-                    </div>
-                  </form>
-                </div>
-              )}
+              {!joinHouse && <CreateHouse />}
             </div>
           </div>
         </div>
